+++ src/imgscanner/ImgScannerImpl.h
@@ -55,7 +55,7 @@ namespace dmscanlib {
             int getScannerCapability();
 
             HANDLE acquireImage(unsigned dpi, int brightness, int contrast,
-                    const ScanRegion<double> & bbox);
+                    const ScanRegion<float> & bbox);
 
             HANDLE acquireFlatbed(unsigned dpi, int brightness, int contrast);
