+++ src/jni/DmScanLibJniInternal.h
@@ -34,14 +34,14 @@ jobject createDecodeResultObject(JNIEnv * env, int resultCode);
 jobject createDecodeResultObject(JNIEnv * env, int resultCode,
         const std::map<std::string, const WellDecoder *> & wellDecoders);
 
-std::unique_ptr<BoundingBox<double> > getBoundingBox(
+std::unique_ptr<BoundingBox<float> > getBoundingBox(
         JNIEnv *env, jobject bboxJavaObj);
 
-std::unique_ptr<ScanRegion<double> > getScanRegion(
+std::unique_ptr<ScanRegion<float> > getScanRegion(
         JNIEnv *env, jobject regionJavaObj);
 
 int getWellRectangles(JNIEnv *env, jsize numWells, jobjectArray _wellRects,
-        std::vector<std::unique_ptr<const WellRectangle<double> > > & wellRects);
+        std::vector<std::unique_ptr<const WellRectangle<float> > > & wellRects);
 
 } /* namespace */
