+++ src/jni/DmScanLibJniWin32.cpp
@@ -108,7 +108,7 @@ JNIEXPORT jobject JNICALL Java_edu_ualberta_med_scannerconfig_dmscanlib_ScanLib_
     unsigned brightness = static_cast<unsigned>(_brightness);
     unsigned contrast = static_cast<unsigned>(_contrast);
     const char *filename = env->GetStringUTFChars(_filename, 0);
-    std::unique_ptr<dmscanlib::ScanRegion<double> > bbox = 
+    std::unique_ptr<dmscanlib::ScanRegion<float> > bbox = 
 		dmscanlib::jni::getScanRegion(env, _region);
 
     dmscanlib::DmScanLib dmScanLib(verbose);
@@ -162,11 +162,11 @@ JNIEXPORT jobject JNICALL Java_edu_ualberta_med_scannerconfig_dmscanlib_ScanLib_
     unsigned dpi = static_cast<unsigned>(_dpi);
     unsigned brightness = static_cast<unsigned>(_brightness);
     unsigned contrast = static_cast<unsigned>(_contrast);
-    std::vector<std::unique_ptr<const dmscanlib::WellRectangle<double> > > wellRects;
+    std::vector<std::unique_ptr<const dmscanlib::WellRectangle<float> > > wellRects;
 
     std::unique_ptr<dmscanlib::DecodeOptions> decodeOptions = 
 		dmscanlib::DecodeOptions::getDecodeOptionsViaJni(env, _decodeOptions);
-    std::unique_ptr<dmscanlib::ScanRegion<double> > scanRegion = 
+    std::unique_ptr<dmscanlib::ScanRegion<float> > scanRegion = 
 		dmscanlib::jni::getScanRegion(env, _region);
 
 	jsize numWells = env->GetArrayLength(_wellRects);
