+++ src/test/TestCommon.h
@@ -27,17 +27,17 @@ void getWellRectsForBoundingBox(
         const BoundingBox<unsigned> & bbox,
         const unsigned rows,
         const unsigned cols,
-        std::vector<std::unique_ptr<const WellRectangle<double> > > & wellRects);
+        std::vector<std::unique_ptr<const WellRectangle<float> > > & wellRects);
 
 std::unique_ptr<DecodeOptions> getDefaultDecodeOptions();
 
 int decodeImage(std::string fname, DmScanLib & dmScanLib, unsigned rows, unsigned cols);
 
-std::unique_ptr<const BoundingBox<double>> getWellsBoundingBox(
-        const BoundingBox<double> & bbox);
+std::unique_ptr<const BoundingBox<float>> getWellsBoundingBox(
+        const BoundingBox<float> & bbox);
 
-std::unique_ptr<const ScanRegion<double>> getWiaBoundingBox(
-        const ScanRegion<double> & bbox);
+std::unique_ptr<const ScanRegion<float>> getWiaBoundingBox(
+        const ScanRegion<float> & bbox);
 
 } /* namespace */
